requirements.txt
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
et-xmlfile==1.1.0
idna==3.4
lml==0.1.0
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
psycopg2==2.9.5
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
texttable==1.6.7
urllib3==1.26.14
